Improvements for streamlit server
1. Finally add caching to `Client.interface`
2. Move streamlit form generation from template
3. Add spinner after request is made
4. TODO: easier config for alternative template
